+++ openmetadata-ui/src/main/resources/ui/src/auth-provider-old/AuthProvider.interface.ts
@@ -22,7 +22,7 @@ export type UserProfile = {
   email: string;
   name: string;
   picture: string;
-  locale?: string;
+  locale: string;
 };
 
 export type OidcUser = {
@@ -30,8 +30,3 @@ export type OidcUser = {
   scope: string;
   profile: UserProfile;
 };
-
-export interface AuthenticatorRef {
-  invokeLogin: () => void;
-  invokeLogout: () => void;
-}
