+++ openmetadata-ui/src/main/resources/ui/src/auth-provider/refreshToken.js
@@ -1,3 +1,16 @@
+/*
+ *  Copyright 2021 Collate
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
 import { oidcTokenKey } from '../constants/constants';
 
 export const refreshTokenSetup = (res) => {
@@ -8,9 +21,6 @@ export const refreshTokenSetup = (res) => {
     try {
       const newAuthRes = await res.reloadAuthResponse();
       refreshTiming = (newAuthRes.expires_in || 3600 - 1 * 60) * 1000;
-      // eslint-disable-next-line no-console
-      console.log('newAuthRes:', newAuthRes);
-      // saveUserToken(newAuthRes.access_token);  <-- save new token
       localStorage.setItem(oidcTokenKey, newAuthRes.id_token);
 
       // Setup the other timer after the first one
