+++ openmetadata-ui/src/main/resources/ui/src/authenticators/MsalAuthenticator.tsx
@@ -1,9 +1,21 @@
+/*
+ *  Copyright 2021 Collate
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
 import {
   InteractionRequiredAuthError,
   InteractionStatus,
 } from '@azure/msal-browser';
 import { useAccount, useIsAuthenticated, useMsal } from '@azure/msal-react';
-// import PropTypes from 'prop-types';
 import React, {
   forwardRef,
   Fragment,
@@ -11,11 +23,13 @@ import React, {
   useEffect,
   useImperativeHandle,
 } from 'react';
-// import { useHistory } from 'react-router-dom';
+import { useAuthContext } from '../auth-provider/AuthProvider';
+import {
+  AuthenticatorRef,
+  OidcUser,
+} from '../auth-provider/AuthProvider.interface';
 import { oidcTokenKey, ROUTES } from '../constants/constants';
 import { msalLoginRequest } from '../utils/AuthProvider.util';
-import { useAuthContext } from './AuthProviderV1';
-import { AuthenticatorRef, OidcUser } from './AuthProviderV1.interface';
 
 interface Props {
   children: ReactNode;
@@ -81,8 +95,6 @@ const MsalAuthenticator = forwardRef<AuthenticatorRef, Props>(
           .acquireTokenSilent(tokenRequest)
           .then((response) => {
             // Call your API with the access token and return the data you need to save in state
-            // eslint-disable-next-line no-console
-            console.log(response);
             const { idToken, scopes, account } = response;
             const user = {
               // eslint-disable-next-line @typescript-eslint/camelcase
@@ -96,7 +108,6 @@ const MsalAuthenticator = forwardRef<AuthenticatorRef, Props>(
             };
             setIsAuthenticated(isMsalAuthenticated);
             localStorage.setItem(oidcTokenKey, idToken);
-            //   refreshTokenSetup(res);
             onLoginSuccess(user);
           })
           .catch(async (e) => {
