+++ openmetadata-ui/src/main/resources/ui/src/components/app-bar/Appbar.tsx
@@ -19,6 +19,7 @@ import { Match } from 'Models';
 import React, { useEffect, useState } from 'react';
 import { Link, useHistory, useLocation, useRouteMatch } from 'react-router-dom';
 import appState from '../../AppState';
+import { useAuthContext } from '../../auth-provider/AuthProvider';
 import { getVersion } from '../../axiosAPIs/miscAPI';
 import {
   getExplorePathWithSearch,
@@ -28,7 +29,6 @@ import {
 } from '../../constants/constants';
 import { urlGitbookDocs, urlJoinSlack } from '../../constants/url.const';
 import { useAuth } from '../../hooks/authHooks';
-import { userSignOut } from '../../utils/AuthUtils';
 import { addToRecentSearched } from '../../utils/CommonUtils';
 import SVGIcons, { Icons } from '../../utils/SvgUtils';
 import { COOKIE_VERSION } from '../Modals/WhatsNewModal/whatsNewData';
@@ -39,13 +39,14 @@ const cookieStorage = new CookieStorage();
 const Appbar: React.FC = (): JSX.Element => {
   const location = useLocation();
   const history = useHistory();
+  const { isFirstTimeUser } = useAuth(location.pathname);
   const {
-    isAuthenticatedRoute,
-    isSignedIn,
-    isFirstTimeUser,
     isAuthDisabled,
+    isAuthenticated,
+    isProtectedRoute,
     isTourRoute,
-  } = useAuth(location.pathname);
+    onLogoutHandler,
+  } = useAuthContext();
   const match: Match | null = useRouteMatch({
     path: ROUTES.EXPLORE_WITH_SEARCH,
   });
@@ -190,7 +191,7 @@ const Appbar: React.FC = (): JSX.Element => {
       name: 'Logout',
       to: '',
       disabled: false,
-      method: userSignOut,
+      method: onLogoutHandler,
     },
   ];
 
@@ -246,7 +247,9 @@ const Appbar: React.FC = (): JSX.Element => {
 
   return (
     <>
-      {isAuthenticatedRoute && isSignedIn && !isTourRoute ? (
+      {isProtectedRoute(location.pathname) &&
+      (isAuthDisabled || isAuthenticated) &&
+      !isTourRoute(location.pathname) ? (
         <NavBar
           handleFeatureModal={handleFeatureModal}
           handleKeyDown={handleKeyDown}
